--- src/main/java/com/hyunec/cosmicbaseballinit/domain/baseball/model/BattingGenerator.java
@@ -1,5 +1,0 @@
-package com.hyunec.cosmicbaseballinit.domain.baseball.model;
-
-public interface BattingGenerator {
-  Batting generator();
-}
