--- src/main/java/com/hyunec/cosmicbaseballinit/domain/baseball/model/RandomStrategy.java
@@ -1,6 +1,0 @@
-package com.hyunec.cosmicbaseballinit.domain.baseball.model;
-
-public interface RandomStrategy {
-
-  int getNumber();
-}
