+++ src/main/java/com/hyunec/cosmicbaseballinit/domain/baseball/model/utils/generator/BattingGenerator.java
@@ -1,0 +1,7 @@
+package com.hyunec.cosmicbaseballinit.domain.baseball.model.utils.generator;
+
+import com.hyunec.cosmicbaseballinit.domain.baseball.model.Batting;
+
+public interface BattingGenerator {
+  Batting generator();
+}
