+++ src/main/java/com/hyunec/cosmicbaseballinit/domain/baseball/model/utils/radom/RandomStrategy.java
@@ -1,0 +1,6 @@
+package com.hyunec.cosmicbaseballinit.domain.baseball.model.utils.radom;
+
+public interface RandomStrategy {
+
+  int getNumber();
+}
