+++ src/main/java/com/hyunec/cosmicbaseballinit/web/GameController.java
@@ -1,26 +1,32 @@
 package com.hyunec.cosmicbaseballinit.web;
 
-import com.hyunec.cosmicbaseballinit.domain.baseball.model.Batting;
-import com.hyunec.cosmicbaseballinit.domain.baseball.model.BattingGenerator;
 import com.hyunec.cosmicbaseballinit.domain.baseball.model.BattingResult;
 import com.hyunec.cosmicbaseballinit.domain.baseball.model.PlateAppearances;
-import lombok.RequiredArgsConstructor;
+import com.hyunec.cosmicbaseballinit.domain.baseball.model.utils.generator.BattingGenerator;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
-@RequiredArgsConstructor
 @RestController
 public class GameController {
 
-    private final PlateAppearances plateAppearances;
-    private final BattingGenerator battingGenerator;
+  private final PlateAppearances plateAppearances;
+  private final BattingGenerator battingGenerator;
 
-    @GetMapping("/game/batting")
-    public BattingResult batting() {
-        plateAppearances.batting(battingGenerator.generator());
-        log.info("### plateAppearances.result()={}", plateAppearances.result());
-        return plateAppearances.result();
-    }
+  public GameController(
+      final PlateAppearances plateAppearances,
+      @Qualifier("randomBattingGenerator")
+      final BattingGenerator battingGenerator) {
+    this.plateAppearances = plateAppearances;
+    this.battingGenerator = battingGenerator;
+  }
+
+  @GetMapping("/game/batting")
+  public BattingResult batting() {
+    plateAppearances.batting(battingGenerator.generator());
+    log.info("### plateAppearances.result()={}", plateAppearances.result());
+    return plateAppearances.result();
+  }
 }
