+++ src/test/java/com/hyunec/cosmicbaseballinit/acceptancetest/NormalBaseballLv1Test.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 
 import com.hyunec.cosmicbaseballinit.domain.baseball.model.Batting;
-import com.hyunec.cosmicbaseballinit.domain.baseball.model.BattingGenerator;
-import com.hyunec.cosmicbaseballinit.domain.baseball.model.RandomBattingGenerator;
-import com.hyunec.cosmicbaseballinit.domain.baseball.model.RandomStrategy;
+import com.hyunec.cosmicbaseballinit.domain.baseball.model.utils.generator.BattingGenerator;
+import com.hyunec.cosmicbaseballinit.domain.baseball.model.utils.generator.RandomBattingGenerator;
+import com.hyunec.cosmicbaseballinit.domain.baseball.model.utils.radom.RandomStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
